% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Wrapper for latent profile analysis to obtain TECH11 and TECH14 outputs}
\usage{
mm(classes, data, usevar = NULL, TECH14 = FALSE, control = mmControl())
}
\arguments{
\item{classes}{the range of latent classes to be modeled}

\item{data}{the data object to be used in the analysis}

\item{usevar}{the variables in the data object to be used in the analysis}

\item{TECH14}{logical, is TECH14 BLRT output required}

\item{control}{Mplus estimation control parameters}
}
\value{

}
\description{
Wrapper for latent profile analysis to obtain TECH11 and TECH14 outputs
}
\examples{
mm(class = 2:3, data = iris, usevar = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))

}
