% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmCompile.R
\name{mmCompile}
\alias{mmCompile}
\title{Compile Mplus LPA output}
\usage{
mmCompile(path = NULL)
}
\arguments{
\item{path}{working directory of the Mplus LPA output files to be compiled}
}
\value{

}
\description{
Compile Mplus LPA output
}
\examples{
mmCompile(path  = getwd())
}
