% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmControl.R
\name{mmControl}
\alias{mmControl}
\title{Estimation specification for Mplus}
\usage{
mmControl(
  processors = 2,
  stiterations = 50,
  starts_initial = 100,
  starts_final = 20,
  LL_replication = 5,
  max_iteration = 5,
  lrtboot = 100,
  lrtstarts_initial = 100,
  lrtstarts_final = 20
)
}
\arguments{
\item{processors}{number of processors}

\item{stiterations}{number of initial stage iterations}

\item{starts_initial}{number of initial stage starts}

\item{starts_final}{number of final stage optimizations}

\item{LL_replication}{minimal number of replications for model log-likelihood}

\item{max_iteration}{maximum number of iterations for finding the optimal seed}

\item{lrtboot}{number of bootstrap draws for TECH14}

\item{lrtstarts_initial}{number of initial stage starts for TECH14}

\item{lrtstarts_final}{number of final stage optimizations for TECH14}
}
\value{

}
\description{
Estimation specification for Mplus
}
\examples{
mmControl()
}
